+++ src/main/java/chess/ConsoleApplication.java
@@ -1,20 +1,21 @@
 package chess;
 
 import chess.controller.console.ConsoleController;
-import chess.dao.MysqlChessDao;
-import chess.dao.StaticMemoryChessDao;
+import chess.mysql.ChessGameManagerRepositoryImpl;
+import chess.mysql.dao.MysqlChessDao;
+import chess.mysql.dao.StaticMemoryChessDao;
 import chess.service.ChessServiceImpl;
 
 import java.util.Objects;
 
-import static chess.dao.ChessConnection.getConnection;
+import static chess.mysql.dao.ChessConnection.getConnection;
 
 public class ConsoleApplication {
     public static void main(String[] args) {
-        ChessServiceImpl chessServiceImpl = new ChessServiceImpl(new StaticMemoryChessDao());
+        ChessServiceImpl chessServiceImpl = new ChessServiceImpl(new ChessGameManagerRepositoryImpl(new StaticMemoryChessDao()));
 
         if (!Objects.isNull(getConnection())) {
-            chessServiceImpl = new ChessServiceImpl(new MysqlChessDao());
+            chessServiceImpl = new ChessServiceImpl(new ChessGameManagerRepositoryImpl(new MysqlChessDao()));
         }
 
         ConsoleController chessController = new ConsoleController(chessServiceImpl);
