+++ src/main/java/chess/mysql/dao/ChessConnection.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.mysql.dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
