+++ src/main/java/chess/mysql/dao/DbConnectionTemplate.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.mysql.dao;
 
 import java.sql.*;
 import java.util.Objects;
