+++ src/test/java/chess/mysql/dao/JdbcTemplateChessDaoTest.java
@@ -1,12 +1,10 @@
-package chess.dao;
+package chess.mysql.dao;
 
-import chess.dao.dto.ChessGame;
-import org.junit.jupiter.api.BeforeEach;
+import chess.mysql.dao.dto.ChessGame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestPropertySource;
