+++ src/test/java/chess/mysql/dao/MysqlChessDaoTest.java
@@ -1,11 +1,10 @@
-package chess.dao;
+package chess.mysql.dao;
 
-import chess.dao.dto.ChessGame;
+import chess.mysql.dao.dto.ChessGame;
 import chess.chessgame.domain.board.Square;
 import chess.chessgame.domain.manager.ChessGameManager;
 import chess.chessgame.domain.manager.ChessGameManagerFactory;
 import chess.chessgame.domain.piece.Pawn;
-import chess.chessgame.domain.piece.attribute.Color;
 import chess.chessgame.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,7 +16,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Optional;
 
-import static chess.chessgame.domain.piece.attribute.Color.BLACK;
 import static chess.chessgame.domain.piece.attribute.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
